Changing the type of gpu content buffer to see what exactly is in it
#30: port aggregations from opencl/cuda implementation

Task-Url: http://github.com/meisam/spark/issues/issue/30

+++ core/src/main/scala/org/apache/spark/rdd/GpuAggregationPartition.scala
@@ -42,7 +42,8 @@ class GpuAggregationPartition[T <: Product: TypeTag](context: OpenCLContext, par
       offsetIndex += 1
     }
     
-    val gpuContentResults = new Array[Int](totalSize)
+    
+    val gpuContentResults = new Array[Byte](totalSize)
     deviceToHostCopy[Byte](gpuContent, Pointer.to(gpuContentResults), totalSize, 0)
     printf("gpuContentResults = %s\n", gpuContentResults.mkString(","))
